feat: refactor TellMorePage to split campaign and newsletter choices
