+++ agir/front/components/authentication/Connexion/TellMore/TellMorePage.js
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useState } from "react";
-import { Redirect, useRouteMatch } from "react-router-dom";
+import { Redirect, useRouteMatch, useLocation } from "react-router-dom";
 
 import { useCustomAnnouncement } from "@agir/activity/common/hooks";
 import { usePush } from "@agir/notifications/push/subscriptions";
@@ -8,15 +8,22 @@ import { useMobileApp } from "@agir/front/app/hooks";
 
 import TellMore from "./TellMore";
 import ChooseCampaign from "./ChooseCampaign";
+import ChooseNewsletters from "./ChooseNewsletters";
 import DeviceNotificationSubscription from "./DeviceNotificationSubscription";
 
 const TellMorePage = () => {
+  const location = useLocation();
+
   const isTellMorePage = useRouteMatch(routeConfig.tellMore.getLink());
 
   const { available, isSubscribed, subscribe, ready, errorMessage } = usePush();
 
   const [hasCampaign, dismissCampaign, campaignIsLoading] =
     useCustomAnnouncement("chooseCampaign");
+
+  const [hasNewsletters, dismissNewsletters, newslettersAreLoading] =
+    useCustomAnnouncement("ChooseNewsletters");
+
   const [hasTellMore, dismissTellMore, tellMoreIsLoading] =
     useCustomAnnouncement("tellMore");
 
@@ -44,16 +51,25 @@ const TellMorePage = () => {
     };
   }, [hasDeviceNotificationSubscription, isMobileApp, ready]);
 
-  if (!isTellMorePage && (hasCampaign || hasTellMore)) {
+  if (!isTellMorePage && (hasCampaign || hasNewsletters || hasTellMore)) {
     return <Redirect to={routeConfig.tellMore.getLink()} />;
   }
 
-  if (campaignIsLoading || tellMoreIsLoading) {
+  if (campaignIsLoading || tellMoreIsLoading || newslettersAreLoading) {
     return null;
   }
 
   if (hasCampaign) {
-    return <ChooseCampaign dismiss={dismissCampaign} />;
+    return (
+      <ChooseCampaign
+        fromSignup={location.hash && location.hash.includes("agir_id")}
+        dismiss={dismissCampaign}
+      />
+    );
+  }
+
+  if (hasNewsletters) {
+    return <ChooseNewsletters dismiss={dismissNewsletters} />;
   }
 
   if (hasTellMore) {
